Bug 1525199 - Part 1 - removed size_t from IPDL messages for Cameras; r=gcp
Differential Revision: https://phabricator.services.mozilla.com/D19191

UltraBlame original commit: cd21500d51113327d995e7e5c053d525ba0bbed9

+++ dom/media/systemservices/VideoFrameUtils.h
@@ -24,7 +24,7 @@ class VideoFrameUtils {
  public:
   
   
-  static size_t TotalRequiredBufferSize(const webrtc::VideoFrame& frame);
+  static uint32_t TotalRequiredBufferSize(const webrtc::VideoFrame& frame);
 
   
   static void InitFrameBufferProperties(
